Commit 552 * Fix a couple of compiler warnings from freebsd-5 * Revamp move_to_prev_word() so its behavior isn't stinky * Same thing with move_to_next_word(), and they need to be commented. * Now move_to_abs_word() doesn't need to skip leading spaces. * Remove now pointless "strip extraneous spaces" feature. * Don't worry about cheating in real_extractw * Fix last_arg() per all of the above

+++ source/functions.c
@@ -1,4 +1,4 @@
-/* $EPIC: functions.c,v 1.128 2003/07/10 13:08:56 jnelson Exp $ */
+/* $EPIC: functions.c,v 1.129 2003/07/14 05:58:27 jnelson Exp $ */
 /*
  * functions.c -- Built-in functions for ircII
  *
@@ -3275,24 +3275,24 @@ BUILT_IN_FUNCTION(function_tdiff2, input)
 
 	if (days)
 	{
-		snprintf(tstr, size, "%ldd ", days);
+		snprintf(tstr, size, "%ldd ", (long)days);
 		size -= strlen(tstr);
 		tstr += strlen(tstr);
 	}
 	if (hours)
 	{
-		snprintf(tstr, size, "%ldh ", hours);
+		snprintf(tstr, size, "%ldh ", (long)hours);
 		size -= strlen(tstr);
 		tstr += strlen(tstr);
 	}
 	if (minutes)
 	{
-		snprintf(tstr, size, "%ldm ", minutes);
+		snprintf(tstr, size, "%ldm ", (long)minutes);
 		size -= strlen(tstr);
 		tstr += strlen(tstr);
 	}
 	if (seconds || (!days && !hours && !minutes))
-		snprintf(tstr, size, "%lds", seconds);
+		snprintf(tstr, size, "%lds", (long)seconds);
 	else
 		*--tstr = 0;	/* chop off that space! */
